Requested changes done except typing.
