+++ public/app/core/angular_wrappers.ts
@@ -6,8 +6,6 @@ import { SearchResult } from './components/search/SearchResult';
 import { TagFilter } from './components/TagFilter/TagFilter';
 import { SideMenu } from './components/sidemenu/SideMenu';
 import AppNotificationList from './components/AppNotifications/AppNotificationList';
-
-import { Teste } from 'app/features/dashboard/custom_time_ranges/reactTest';
 import { CustomTimeRangeEditor } from 'app/features/dashboard/custom_time_ranges/editorReact';
 
 export function registerAngularDirectives() {
@@ -22,6 +20,5 @@ export function registerAngularDirectives() {
     ['onSelect', { watchDepth: 'reference' }],
     ['tagOptions', { watchDepth: 'reference' }],
   ]);
-  react2AngularDirective('teste', Teste, []);
   react2AngularDirective('customtimerangeeditor', CustomTimeRangeEditor, ['dashboard']);
 }
