+++ public/app/features/dashboard/custom_time_ranges/TimeRangeList.tsx
@@ -1,100 +1,29 @@
 import React, { PureComponent } from 'react';
-
-const width1 = {
-  width: '1%',
-};
-
-const lightBlue = {
-  color: '#32d1df',
-};
-
-const green = {
-  color: '#269242',
-};
+import Table from './TimeRangeTable';
 
 export interface Props {
   ranges: any;
-  up: any;
-  down: any;
+  move: any;
   edit: any;
   delete: any;
   add: any;
 }
 
 export default class List extends PureComponent<Props> {
-  constructor(props) {
-    super(props);
-    this.state = { show: false };
-  }
-
-  clickUp(index) {
-    this.props.up(index);
-  }
-  clickDown(index) {
-    this.props.down(index);
-  }
-  clickEdit(index) {
+  move = (index, dir) => {
+    this.props.move(index, dir);
+  };
+  edit = index => {
     this.props.edit(index);
-  }
-  clickDelete(index) {
+  };
+  delete = index => {
     this.props.delete(index);
-  }
+  };
   setupNew = () => {
     this.props.add();
   };
 
   render() {
-    const NewDay = props => {
-      return (
-        <span style={green} className="pointer">
-          Starts a new day
-        </span>
-      );
-    };
-
-    const rowLen = this.props.ranges.length;
-    const tableRows = this.props.ranges.map((range, index) => {
-      return (
-        <tr key={index}>
-          <td className="pointer" onClick={() => this.clickEdit(index)}>
-            <i className="fa fa-clock-o" />
-            <span>&nbsp;&nbsp;shift</span>
-          </td>
-          <td onClick={() => this.clickEdit(index)}>
-            <span className="pointer">
-              <strong style={lightBlue}>Name: </strong>
-              {range.name}
-              <span>&nbsp;&nbsp;</span>
-            </span>
-            <span className="pointer">
-              <strong style={lightBlue}>From: </strong>
-              {range.from}
-              <span>&nbsp;&nbsp;</span>
-            </span>
-            <span className="pointer">
-              <strong style={lightBlue}>To: </strong>
-              {range.to}
-              <span>&nbsp;&nbsp;</span>
-            </span>
-            {range.newDay ? <NewDay /> : null}
-          </td>
-          <td style={width1}>
-            {index === 0 ? null : <i className="pointer fa fa-arrow-up" onClick={() => this.clickUp(index)} />}
-          </td>
-          <td style={width1}>
-            {index === rowLen - 1 ? null : (
-              <i className="pointer fa fa-arrow-down" onClick={() => this.clickDown(index)} />
-            )}
-          </td>
-          <td style={width1}>
-            <a className="btn btn-danger btn-mini">
-              <i className="fa fa-remove" onClick={() => this.clickDelete(index)} />
-            </a>
-          </td>
-        </tr>
-      );
-    });
-
     return (
       <div>
         <div className="page-action-bar">
@@ -103,15 +32,7 @@ export default class List extends PureComponent<Props> {
             <i className="fa fa-plus" /> New
           </a>
         </div>
-        <table className="filter-table filter-table--hover">
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Info</th>
-            </tr>
-          </thead>
-          <tbody>{tableRows}</tbody>
-        </table>
+        <Table ranges={this.props.ranges} move={this.move} edit={this.edit} delete={this.delete} />
       </div>
     );
   }
