--- public/app/features/dashboard/custom_time_ranges/editor.html
@@ -1,126 +1,0 @@
-<h3 class="dashboard-settings__header">
-  <a ng-click="ctrl.backToList()">Custom Time Ranges</a>
-  <span ng-show="ctrl.mode === 'new'">&gt; New</span>
-  <span ng-show="ctrl.mode === 'edit'">&gt; Edit</span>
-</h3>
-
-<div ng-if="ctrl.mode == 'list'">
-  <div ng-if="ctrl.dashboard.ranges.length === 0">
-    <div class="empty-list-cta">
-      <div class="empty-list-cta__title">
-        There are no time ranges added yet
-      </div>
-      <a ng-click="ctrl.setupNew()" class="empty-list-cta__button btn btn-xlarge btn-success">
-        <i class="fa fa-clock-o"></i>
-        Add Time Range
-      </a>
-      <div class="grafana-info-box">
-        <h5>What are Custom Time Ranges?</h5>
-        <p>
-					Custom Time Ranges allow you to define and browse your own ranges easily! Your ranges will be added to timepicker for quick and convenient access.
-        </p>
-      </div>
-    </div>
-  </div>
-
-  <div ng-if="ctrl.dashboard.ranges.length > 0">
-    <div class="page-action-bar">
-      <div class="page-action-bar__spacer"></div>
-      <a type="button" class="btn btn-success" ng-click="ctrl.setupNew()">
-        <i class="fa fa-plus"></i> New</a>
-    </div>
-    <table class="filter-table filter-table--hover">
-      <thead>
-        <tr>
-          <th>Name</th>
-          <th>Info</th>
-          <th colspan="3"></th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr ng-repeat="range in ctrl.dashboard.ranges">
-          <td  ng-click="ctrl.editRange(range)" class="pointer">
-            <i class="fa fa-clock-o"></i>
-            {{range.type}}
-          </td>
-          <td>
-            <span ng-if="range.name" ng-click="ctrl.editRange(range)" class="pointer">
-              <strong ng-style="{'color':'#32d1df'}">Name: </strong>{{range.name}}
-            </span>
-            <span ng-if="range.from" ng-click="ctrl.editRange(range)" class="pointer">
-              <strong ng-style="{'color':'#32d1df'}">From: </strong>{{range.from}}
-            </span>
-            <span ng-if="range.to" ng-click="ctrl.editRange(range)" class="pointer">
-              <strong ng-style="{'color':'#32d1df'}">To: </strong>{{range.to}}
-            </span>
-            <span ng-if="range.newDay" ng-click="ctrl.editRange(range)" class="pointer">
-              <strong ng-style="{'color':'#269242'}">Starts a new day</strong>
-            </span>
-          </td>
-          <td style="width: 1%">
-            <i ng-click="ctrl.moveRange($index, -1)" ng-hide="$first" class="pointer fa fa-arrow-up"></i>
-          </td>
-          <td style="width: 1%">
-            <i ng-click="ctrl.moveRange($index, 1)" ng-hide="$last" class="pointer fa fa-arrow-down"></i>
-          </td>
-          <td style="width: 1%">
-            <a ng-click="ctrl.deleteRange($index)" class="btn btn-danger btn-mini" ng-hide="annotation.builtIn">
-              <i class="fa fa-remove"></i>
-            </a>
-          </td>
-        </tr>
-      </tbody>
-    </table>
-  </div>
-</div>
-
-<form ng-if="ctrl.mode == 'edit' || ctrl.mode == 'new'" name="ctrl.form">
-  <div class="gf-form-group">
-    <div class="gf-form-group">
-      <div class="gf-form">
-        <span class="gf-form-label width-8">Type</span>
-        <div class="gf-form-select-wrapper width-10">
-          <select class="gf-form-input" ng-model="ctrl.range.type" ng-options="f for f in ['shift']"></select>
-        </div>
-      </div>
-    </div>
-
-    <div class="gf-form-group">
-      <div class="gf-form" ng-show="ctrl.range.type === 'shift'">
-        <span class="gf-form-label width-5">Name</span>
-        <input type="text" class="gf-form-input width-10" name="name" placeholder="name" ng-model='ctrl.range.name' required ng-pattern="namePattern">
-      </div>
-      <div class="gf-form" ng-show="ctrl.range.type === 'shift'">
-        <span class="gf-form-label width-5">From</span>
-        <input id="fromHourInput" type="number" min="0" max="23" step="1" class="gf-form-input width-3" name="fromHour" placeholder="08" ng-model='ctrl.range.fromHour' required onblur="if(this.value.toString().length === 1)this.value='0'+this.value;">
-        :&nbsp
-        <input id="fromMinInput" type="number" min="0" max="59" step="1" class="gf-form-input width-3" name="fromMin" placeholder="00" ng-model='ctrl.range.fromMin' required onblur="if(this.value.toString().length === 1)this.value='0'+this.value;">
-      </div>
-      <div class="gf-form" ng-show="ctrl.range.type === 'shift'">
-        <span class="gf-form-label width-5">To</span>
-        <input id="toHourInput" type="number" min="0" max="23" step="1" class="gf-form-input width-3" name="toHour" placeholder="15" ng-model='ctrl.range.toHour' required onblur="if(this.value.toString().length === 1)this.value='0'+this.value;">
-        :&nbsp
-        <input id="toMinInput" type="number" min="0" max="59" step="1" class="gf-form-input width-3" name="toMin" placeholder="30" ng-model='ctrl.range.toMin' required onblur="if(this.value.toString().length === 1)this.value='0'+this.value;">
-      </div>
-    </div>
-
-    <div class="gf-form-group">
-      <div class="gf-form" ng-show="ctrl.range.type === 'shift'">
-          <gf-form-switch class="gf-form"
-					label="Starts a new day"
-					label-class="width-12"
-					tooltip="Check when time range crosses midnight"
-					checked="ctrl.range.newDay">
-				</gf-form-switch>
-      </div>
-    </div>
-
-  </div>
-  <button ng-show="ctrl.range.type === 'shift'" class="btn btn-success" ng-if="ctrl.mode == 'new'" ng-click="ctrl.addRange()">
-    Add
-  </button>
-  <button class="btn btn-success" ng-if="ctrl.mode == 'edit'" ng-click="ctrl.saveRange()">
-    Update
-  </button>
-</form>
-<div id="teste"> React Tests Bellow</div>
