+++ public/app/features/dashboard/custom_time_ranges/editorReact.tsx
@@ -1,114 +1,121 @@
 import React, { PureComponent } from 'react';
+import List from './TimeRangeList';
+import NoTimeRanges from './NoTimeRanges';
+import TimeRangeEditor from './TimeRangeEditor';
+import { StoreState } from 'app/types';
+import { hot } from 'react-hot-loader';
+import { connect } from 'react-redux';
 import { DashboardModel } from '../dashboard_model';
 
-export interface Props {}
-export interface State {
-  value: string;
+export function timeToString(hour, minute) {
+  if (hour % 1 === 0 && minute % 1 === 0 && hour < 24 && hour >= 0 && minute < 60 && minute >= 0) {
+    if (hour.toString().length === 1) {
+      hour = '0' + hour;
+    }
+    if (minute.toString().length === 1) {
+      minute = '0' + minute;
+    }
+    const TimeString = hour + ':' + minute;
+    return TimeString;
+  }
+  throw new Error('Invalid input');
+}
+
+export interface Props {
   dashboard: DashboardModel;
 }
 
-export class CustomTimeRangeEditor extends PureComponent<Props, State> {
+const ranges = [
+  {
+    name: 'time range 1',
+    from: '08:00',
+    to: '16:00',
+    newDay: false,
+  },
+  {
+    name: 'time range 2',
+    from: '16:00',
+    to: '00:00',
+    newDay: true,
+  },
+  {
+    name: 'time range 3',
+    from: '00:00',
+    to: '08:00',
+    newDay: false,
+  },
+];
+
+let i;
+
+export class CustomTimeRangeEditor extends PureComponent<Props, { mode: string }> {
   constructor(props) {
     super(props);
-    this.state = null;
-    this.removeButton = this.removeButton.bind(this);
-  }
 
-  addHandler() {
-    console.log('Add');
+    this.state = {
+      mode: 'list',
+    };
+    this.setupNew = this.setupNew.bind(this);
+    this.add = this.add.bind(this);
+    this.edit = this.edit.bind(this);
+    this.saveRange = this.saveRange.bind(this);
+    this.backToList = this.backToList.bind(this);
+    console.log('here', this.props);
   }
-  updateHandler() {
-    console.log('Update');
+  add(range) {
+    this.setState({ mode: 'list' });
+    console.log('Add clicked', range);
   }
-  handleChange(event) {
-    this.setState({ value: event.target.value });
+  delete(index) {
+    i = index;
+    console.log(index, 'Delete clicked');
   }
-  removeButton() {
-    console.log('Remove');
-    const dasboard = this.state.dashboard;
-    dasboard.test();
+  edit(index) {
+    i = index;
+    this.setState({ mode: 'edit' });
+    console.log(index, 'Edit clicked');
+  }
+  move(index, dir) {
+    i = index;
+    console.log(index, dir, 'Move clicked');
+  }
+  setupNew() {
+    this.setState({ mode: 'new' });
+    console.log('Setup new clicked');
+  }
+  saveRange(range) {
+    this.setState({ mode: 'list' });
+    console.log('Save range clicked', range);
+  }
+  backToList() {
+    this.setState({ mode: 'list' });
+    console.log('Back to list clicked');
   }
   render() {
+    if (ranges.length === 0 && this.state.mode !== 'new') {
+      this.setState({ mode: 'empty' });
+    }
     return (
-      <form name="form">
-        <div className="gf-form-group">
-          <div className="gf-form">
-            <span className="gf-form-label width-5">Name</span>
-            <input
-              type="text"
-              className="gf-form-input width-10"
-              name="name"
-              placeholder="name"
-              ng-model="ctrl.range.name"
-              required
-            />
-          </div>
-          <div className="gf-form">
-            <span className="gf-form-label width-5">From</span>
-            <input
-              id="fromHourInput"
-              type="number"
-              min="0"
-              max="23"
-              step="1"
-              className="gf-form-input width-3"
-              name="fromHour"
-              placeholder="08"
-              required
-            />
-            <span>&nbsp;:&nbsp;</span>
-            <input
-              id="fromMinInput"
-              type="number"
-              min="0"
-              max="59"
-              step="1"
-              className="gf-form-input width-3"
-              name="fromMin"
-              placeholder="00"
-              required
-            />
-          </div>
-          <div className="gf-form">
-            <span className="gf-form-label width-5">To</span>
-            <input
-              id="toHourInput"
-              type="number"
-              min="0"
-              max="23"
-              step="1"
-              className="gf-form-input width-3"
-              name="toHour"
-              placeholder="15"
-              required
-            />
-            <span>&nbsp;:&nbsp;</span>
-            <input
-              id="toMinInput"
-              type="number"
-              min="0"
-              max="59"
-              step="1"
-              className="gf-form-input width-3"
-              name="toMin"
-              placeholder="30"
-              required
-            />
-          </div>
-        </div>
-        <div className="gf-form-group">
-          <div className="gf-form">
-            <span className="gf-form-label width-10">Starts a new day</span>
-            <input type="checkbox" value="newDay" className="gf-form-switch" />
-          </div>
-        </div>
-        <button className="btn btn-success" onClick={this.addHandler}>
-          Add
-        </button>
-        <button className="btn btn-success" onClick={this.updateHandler}>
-          Update
-        </button>
-      </form>
+      <div>
+        <h3 className="dashboard-settings__header">
+          <a onClick={this.backToList}>Custom Time Ranges</a>
+          {this.state.mode === 'new' ? <span> &gt; New</span> : null}
+          {this.state.mode === 'edit' ? <span> &gt; Edit</span> : null}
+        </h3>
+        {this.state.mode === 'empty' ? <NoTimeRanges add={this.setupNew} /> : null}
+        {this.state.mode === 'list' ? (
+          <List ranges={ranges} add={this.setupNew} move={this.move} edit={this.edit} delete={this.delete} />
+        ) : null}
+        {this.state.mode === 'new' || this.state.mode === 'edit' ? (
+          <TimeRangeEditor range={ranges[i]} add={this.add} update={this.saveRange} mode={this.state.mode} />
+        ) : null}
+      </div>
     );
   }
 }
+
+const mapStateToProps = (state: StoreState) => ({
+  dashboard: state.dashboard,
+});
+
+export default hot(module)(connect(mapStateToProps)(CustomTimeRangeEditor));
