--- public/app/features/dashboard/custom_time_ranges/module.ts
@@ -1,19 +1,0 @@
-import angular from 'angular';
-import _ from 'lodash';
-import { iconMap } from './editor';
-
-export class CustomTimeRangesCtrl {
-  /** @ngInject */
-  constructor($scope, $rootScope, $q, rangeDef) {
-    if (rangeDef.type === 'range') {
-      return $q.when([
-        {
-          icon: iconMap[rangeDef.icon],
-        },
-      ]);
-    }
-
-    return $q.when([]);
-  }
-}
-angular.module('grafana.directives').controller('CustomTimeRangesCtrl', CustomTimeRangesCtrl);
