+++ public/app/features/dashboard/custom_time_ranges/range_model.ts
@@ -1,0 +1,10 @@
+export default class RangeModel {
+  name: string;
+  fromHour?: number;
+  fromMin?: number;
+  toHour?: number;
+  toMin?: number;
+  from?: string;
+  to?: string;
+  newDay?: boolean;
+}
