+++ public/app/features/dashboard/custom_time_ranges/specs/editor.test.ts
@@ -1,4 +1,4 @@
-import { timeToString } from '../editor';
+import { timeToString } from '../editorReact';
 
 describe('CustomTimeRangeEditorCtrl', () => {
   describe('timeToString function', () => {
