+++ public/app/features/dashboard/settings/settings.html
@@ -80,8 +80,7 @@
 </div>
 
 <div class="dashboard-settings__content" ng-if="ctrl.viewId === 'customTimeRanges'">
-	<custom-range-editor dashboard="ctrl.dashboard"></custom-range-editor>
-	<teste name="teste" dashboard="ctrl.dashboard"/>
+	<customtimerangeeditor name="customtimerangeeditor"/>
 </div>
 
 <div class="dashboard-settings__content" ng-if="ctrl.viewId === 'links'" >
